ch20: nicify output

+++ chapter20 - training/train.py
@@ -95,9 +95,9 @@ def make_a_prediction(message, model, valX_enc, valX_dec, valY_dec):
     print(message)
     # preds = model.predict((valX_enc[0:1], valX_dec[0:1]), verbose=0)
     preds = model((valX_enc[0:3], valX_dec[0:3]), training=False)
-    # print(valX_enc[0:1].numpy())
-    # print(valX_dec[0:1].numpy())
-    print(valY_dec[0:3].numpy())
+    # print(tf.cast(valX_enc[0:1], dtype=tf.int32).numpy())
+    # print(tf.cast(valX_dec[0:1], dtype=tf.int32).numpy())
+    print(tf.cast(valY_dec[0:3], dtype=tf.int32).numpy())
     print(tf.argmax(preds, axis=-1))
